forge formatted

+++ contracts/token/EPNS.sol
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.20;
+
 import { BaseHelper } from "../libraries/BaseHelper.sol";
 
 contract EPNS {
@@ -14,25 +15,25 @@ contract EPNS {
     uint8 public constant decimals = 18;
 
     /// @notice Total number of tokens in circulation
-    uint public totalSupply = 100_000_000e18; // 100 million PUSH
+    uint256 public totalSupply = 100_000_000e18; // 100 million PUSH
 
     /// @notice block number when tokens came into circulation
-    uint public born;
+    uint256 public born;
 
     /// @dev Allowance amounts on behalf of others
-    mapping (address => mapping (address => uint96)) internal allowances;
+    mapping(address => mapping(address => uint96)) internal allowances;
 
     /// @dev Official record of token balances for each account
-    mapping (address => uint96) internal balances;
+    mapping(address => uint96) internal balances;
 
     /// @notice Official record of the token block information for the holder
-    mapping (address => uint) public holderWeight;
+    mapping(address => uint256) public holderWeight;
 
     /// @notice Allows holder weight to be reset by other adddresses
-    mapping (address => mapping(address => bool)) public holderDelegation;
+    mapping(address => mapping(address => bool)) public holderDelegation;
 
     /// @notice A record of each accounts delegate
-    mapping (address => address) public delegates;
+    mapping(address => address) public delegates;
 
     /// @notice A checkpoint for marking number of votes from a given block
     struct Checkpoint {
@@ -41,31 +42,34 @@ contract EPNS {
     }
 
     /// @notice A record of votes checkpoints for each account, by index
-    mapping (address => mapping (uint32 => Checkpoint)) public checkpoints;
+    mapping(address => mapping(uint32 => Checkpoint)) public checkpoints;
 
     /// @notice The number of checkpoints for each account
-    mapping (address => uint32) public numCheckpoints;
+    mapping(address => uint32) public numCheckpoints;
 
     /// @notice The EIP-712 typehash for the contract's domain
-    bytes32 public constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
+    bytes32 public constant DOMAIN_TYPEHASH =
+        keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
 
     /// @notice The EIP-712 typehash for the delegation struct used by the contract
-    bytes32 public constant DELEGATION_TYPEHASH = keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
+    bytes32 public constant DELEGATION_TYPEHASH =
+        keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
 
     /// @notice The EIP-712 typehash for the permit struct used by the contract
-    bytes32 public constant PERMIT_TYPEHASH = keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
+    bytes32 public constant PERMIT_TYPEHASH =
+        keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)");
 
     /// @notice A record of states for signing / validating signatures
-    mapping (address => uint) public nonces;
+    mapping(address => uint256) public nonces;
 
     /// @notice An event thats emitted when an account changes its total token weight
-    event HolderWeightChanged(address indexed holder, uint256 amount, uint weight);
+    event HolderWeightChanged(address indexed holder, uint256 amount, uint256 weight);
 
     /// @notice An event thats emitted when an account changes its delegate
     event DelegateChanged(address indexed delegator, address indexed fromDelegate, address indexed toDelegate);
 
     /// @notice An event thats emitted when a delegate account's vote balance changes
-    event DelegateVotesChanged(address indexed delegate, uint previousBalance, uint newBalance);
+    event DelegateVotesChanged(address indexed delegate, uint256 previousBalance, uint256 newBalance);
 
     /// @notice The standard EIP-20 transfer event
     event Transfer(address indexed from, address indexed to, uint256 amount);
@@ -92,7 +96,7 @@ contract EPNS {
      * @param spender The address of the account spending the funds
      * @return The number of tokens approved
      */
-    function allowance(address account, address spender) external view returns (uint) {
+    function allowance(address account, address spender) external view returns (uint256) {
         return allowances[account][spender];
     }
 
@@ -129,7 +133,17 @@ contract EPNS {
      * @param r Half of the ECDSA signature pair
      * @param s Half of the ECDSA signature pair
      */
-    function permit(address owner, address spender, uint rawAmount, uint deadline, uint8 v, bytes32 r, bytes32 s) external {
+    function permit(
+        address owner,
+        address spender,
+        uint256 rawAmount,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    )
+        external
+    {
         uint96 amount;
         if (rawAmount == type(uint256).max) {
             amount = type(uint96).max;
@@ -138,9 +152,10 @@ contract EPNS {
             amount = uint96(rawAmount);
         }
 
-
-        bytes32 domainSeparator = keccak256(abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), BaseHelper.getChainId(), address(this)));
-        bytes32 structHash = keccak256(abi.encode(PERMIT_TYPEHASH, owner, spender, rawAmount, nonces[owner]++, deadline));
+        bytes32 domainSeparator =
+            keccak256(abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), BaseHelper.getChainId(), address(this)));
+        bytes32 structHash =
+            keccak256(abi.encode(PERMIT_TYPEHASH, owner, spender, rawAmount, nonces[owner]++, deadline));
         bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
         address signatory = ecrecover(digest, v, r, s);
 
@@ -158,7 +173,7 @@ contract EPNS {
      * @param account The address of the account to get the balance of
      * @return The number of tokens held
      */
-    function balanceOf(address account) external view returns (uint) {
+    function balanceOf(address account) external view returns (uint256) {
         return balances[account];
     }
 
@@ -168,7 +183,7 @@ contract EPNS {
      * @param rawAmount The number of tokens to transfer
      * @return Whether or not the transfer succeeded
      */
-    function transfer(address dst, uint rawAmount) external returns (bool) {
+    function transfer(address dst, uint256 rawAmount) external returns (bool) {
         uint96 amount = safe96(rawAmount, "Push::transfer: amount exceeds 96 bits");
         _transferTokens(msg.sender, dst, amount);
         return true;
@@ -181,13 +196,14 @@ contract EPNS {
      * @param rawAmount The number of tokens to transfer
      * @return Whether or not the transfer succeeded
      */
-    function transferFrom(address src, address dst, uint rawAmount) external returns (bool) {
+    function transferFrom(address src, address dst, uint256 rawAmount) external returns (bool) {
         address spender = msg.sender;
         uint96 spenderAllowance = allowances[src][spender];
         uint96 amount = safe96(rawAmount, "Push::approve: amount exceeds 96 bits");
 
         if (spender != src && spenderAllowance != type(uint96).max) {
-            uint96 newAllowance = sub96(spenderAllowance, amount, "Push::transferFrom: transfer amount exceeds spender allowance");
+            uint96 newAllowance =
+                sub96(spenderAllowance, amount, "Push::transferFrom: transfer amount exceeds spender allowance");
             allowances[src][spender] = newAllowance;
 
             emit Approval(src, spender, newAllowance);
@@ -200,8 +216,14 @@ contract EPNS {
     /**
      * @notice Return holder units
      */
-    function returnHolderUnits(address account, uint atBlock) external view returns (uint) {
-        return mul256(balances[account], sub256(atBlock, holderWeight[account], "Push::returnHolderUnits: atBlock should be greater than holderWeight"), "Push::returnHolderUnits: ratio exceeds max range");
+    function returnHolderUnits(address account, uint256 atBlock) external view returns (uint256) {
+        return mul256(
+            balances[account],
+            sub256(
+                atBlock, holderWeight[account], "Push::returnHolderUnits: atBlock should be greater than holderWeight"
+            ),
+            "Push::returnHolderUnits: ratio exceeds max range"
+        );
     }
 
     /**
@@ -222,16 +244,19 @@ contract EPNS {
      * @notice Reset holder weight to current block
      */
     function resetHolderWeight(address holder) external {
-        require(holderDelegation[holder][msg.sender] == true || holder == msg.sender, "Push::resetHolderWeight: unauthorized");
+        require(
+            holderDelegation[holder][msg.sender] == true || holder == msg.sender,
+            "Push::resetHolderWeight: unauthorized"
+        );
         holderWeight[holder] = block.number;
 
         emit HolderWeightChanged(holder, balances[holder], block.number);
     }
 
-     /**
-      * @notice Destory `RawAmount` of tokens from a holder `account`
-      * @param rawAmount The number of tokens
-      */
+    /**
+     * @notice Destory `RawAmount` of tokens from a holder `account`
+     * @param rawAmount The number of tokens
+     */
     function burn(uint256 rawAmount) external {
         address account = msg.sender;
 
@@ -260,8 +285,9 @@ contract EPNS {
      * @param r Half of the ECDSA signature pair
      * @param s Half of the ECDSA signature pair
      */
-    function delegateBySig(address delegatee, uint nonce, uint expiry, uint8 v, bytes32 r, bytes32 s) public {
-        bytes32 domainSeparator = keccak256(abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), BaseHelper.getChainId(), address(this)));
+    function delegateBySig(address delegatee, uint256 nonce, uint256 expiry, uint8 v, bytes32 r, bytes32 s) public {
+        bytes32 domainSeparator =
+            keccak256(abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name)), BaseHelper.getChainId(), address(this)));
         bytes32 structHash = keccak256(abi.encode(DELEGATION_TYPEHASH, delegatee, nonce, expiry));
         bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
         address signatory = ecrecover(digest, v, r, s);
@@ -288,7 +314,7 @@ contract EPNS {
      * @param blockNumber The block number to get the vote balance at
      * @return The number of votes the account had as of the given block
      */
-    function getPriorVotes(address account, uint blockNumber) public view returns (uint96) {
+    function getPriorVotes(address account, uint256 blockNumber) public view returns (uint96) {
         require(blockNumber < block.number, "Push::getPriorVotes: not yet determined");
 
         uint32 nCheckpoints = numCheckpoints[account];
@@ -354,18 +380,21 @@ contract EPNS {
         uint96 balance = balances[dst];
         if (balance == 0) {
             holderWeight[dst] = holderWeight[src];
-        }
-        else {
-            uint256 dstWeight = mul256(holderWeight[dst], balance, "Push::_adjustHolderWeight: holder dst weight exceeded limit");
-            uint256 srcWeight = mul256(holderWeight[src], amount, "Push::_adjustHolderWeight: holder src weight exceeded limit");
+        } else {
+            uint256 dstWeight =
+                mul256(holderWeight[dst], balance, "Push::_adjustHolderWeight: holder dst weight exceeded limit");
+            uint256 srcWeight =
+                mul256(holderWeight[src], amount, "Push::_adjustHolderWeight: holder src weight exceeded limit");
 
             uint256 totalWeight = add256(dstWeight, srcWeight, "Push::_adjustHolderWeight: total weight exceeded limit");
             uint256 totalAmount = add256(balance, amount, "Push::_adjustHolderWeight: total amount exceeded limit");
 
             uint256 totalAmountBy2 = div256(totalAmount, 2, "");
-            uint256 roundUpWeight = add256(totalWeight, totalAmountBy2, "Push::_adjustHolderWeight: round up amount exceeded limit");
+            uint256 roundUpWeight =
+                add256(totalWeight, totalAmountBy2, "Push::_adjustHolderWeight: round up amount exceeded limit");
 
-            holderWeight[dst] = div256(roundUpWeight, totalAmount, "Push::_adjustHolderWeight: adjusted holder negative divide");
+            holderWeight[dst] =
+                div256(roundUpWeight, totalAmount, "Push::_adjustHolderWeight: adjusted holder negative divide");
         }
     }
 
@@ -400,13 +429,13 @@ contract EPNS {
         emit DelegateVotesChanged(delegatee, oldVotes, newVotes);
     }
 
-    function safe32(uint n, string memory errorMessage) internal pure returns (uint32) {
-        require(n < 2**32, errorMessage);
+    function safe32(uint256 n, string memory errorMessage) internal pure returns (uint32) {
+        require(n < 2 ** 32, errorMessage);
         return uint32(n);
     }
 
-    function safe96(uint n, string memory errorMessage) internal pure returns (uint96) {
-        require(n < 2**96, errorMessage);
+    function safe96(uint256 n, string memory errorMessage) internal pure returns (uint96) {
+        require(n < 2 ** 96, errorMessage);
         return uint96(n);
     }
